requirements.txt
accelerate==1.6.0
cloudscraper==1.2.71
fastparquet==2024.11.0
faiss-cpu==1.11.0
ipython==9.1.0
numpy==2.2.5
pandas==2.2.3
sentence-transformers==4.1.0
streamlit==1.45.0
tiktoken==0.9.0
tqdm==4.67.1
transformers==4.51.3